+++ src/engraving/libmscore/tuplet.h
@@ -81,7 +81,7 @@ public:
     // Score Tree functions
     EngravingObject* scanParent() const override;
     EngravingObject* scanChild(int idx) const override;
-    int scanChildCount() const override;
+    size_t scanChildCount() const override;
 
     Tuplet* clone() const override { return new Tuplet(*this); }
     void setTrack(int val) override;
